Survivor indices as int: population up to 65536
Survivor indices and ranks were short, which capped the population at 8192
(N = 2P below 32768). They are int now, in the launchers, the kernels, the
survival workspace and the solver, and kMaxPopulation goes up to 65536.

Measured on an RTX 2060 (GPU time of 10 generations of one run):

  KC10-2fl-1rl  71 ms (P=8192)  167 ms (16384)  353 ms (32768)  1183 ms (65536)
  KC30-3fl-1rl 206 ms (P=8192)  439 ms (16384) 1049 ms (32768)  2800 ms (65536)

A complete run of 70 generations with P = 65536 costs 6.7 s of GPU time on
KC10-2fl-1rl and 17.7 s on KC30-3fl-1rl, so from here on the limit is the
O(N^2) dominance counting, not the memory or the index type.

Validation:
- Results stay byte-identical to the previous binary on KC10, KC20 and KC30
  with P = 64, 256, 1024 and 4096.
- Two new tests with P = 32768 and P = 65536 check the invariants of the
  selection (distinct survivors, ordered by rank and crowding, no dominated
  rank-1 survivor); test_kernels takes --quick to skip them under
  compute-sanitizer.
- 22/22 tests pass; memcheck, initcheck, synccheck and racecheck report no
  errors; --verify is OK on the 23 .dat instances with P = 8192 and on
  KC10/KC30 with P = 65536.

+++ include/kernels.cuh
@@ -7,7 +7,7 @@
  * Memory layout (R independent runs, population P, n facilities, OBJ objectives):
  *   genes    short        [R][2P][n]   rows [0, P) survivors, rows [P, 2P) offspring
  *   fitness  unsigned int [R][2P][OBJ]
- *   survivor arrays (index/rank/crowding)  [R][P]
+ *   survivor arrays  index/rank int, crowding float  [R][P]
  *   rng      RngState     [R][2P]
  *
  * Copyright (C) 2019-2026 Andres Pupiales Arevalo <apupiales@gmail.com>
@@ -57,13 +57,13 @@ void launchFitness(const short* genes, unsigned int* fitness, const int* flow, c
 // workspace says so (P > kSingleBlockMaxPopulation, or forced).
 template <int OBJ>
 void launchSurvival(const unsigned int* fitness, int population, int runs,
-                    short* survivorIndex, short* survivorRank, float* survivorCrowding,
+                    int* survivorIndex, int* survivorRank, float* survivorCrowding,
                     SurvivalWorkspace* workspace = nullptr);
 
 // Multi-block NSGA-II survival (any population up to kMaxPopulation); see nsga2_multiblock.cu.
 template <int OBJ>
 void launchSurvivalMultiblock(const unsigned int* fitness, int population, int runs,
-                              short* survivorIndex, short* survivorRank, float* survivorCrowding,
+                              int* survivorIndex, int* survivorRank, float* survivorCrowding,
                               SurvivalWorkspace& workspace);
 
 // Builds the next population of each run:
@@ -73,7 +73,7 @@ void launchSurvivalMultiblock(const unsigned int* fitness, int population, int r
 template <int OBJ>
 void launchReproduce(const short* genes, const unsigned int* fitness,
                      short* nextGenes, unsigned int* nextFitness,
-                     const short* survivorIndex, const short* survivorRank, const float* survivorCrowding,
+                     const int* survivorIndex, const int* survivorRank, const float* survivorCrowding,
                      RngState* rng, int* greedyType, int population, int n, int runs);
 
 // Adapted greedy 2-opt on rows [P, 2P) of each run with O(n) delta evaluation. Leaves the
